feat: explore-with-me

+++ server/src/main/java/ru/practicum/explorewithme/event/location/LocationDto.java
@@ -10,6 +10,6 @@ import javax.persistence.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class LocationDto {
-    private Double lat;
-    private Double lon;
+    private Float lat;
+    private Float lon;
 }
